refactor(path-bar): streamline breadcrumb handling and integrate path bar components
- Moved breadcrumb-related functionality to a new PathBar module for better organization.
- Updated path bar integration in PanelView and related classes to utilize the new PathBar API.
- Simplified breadcrumb rendering logic and improved context menu handling for path bar actions.
- Removed redundant code and ensured consistent usage of path context across components.
