+++ Source/NimbleCommander/NimbleCommander/States/FilePanels/NCPanelPathBarPresentation.mm
@@ -1,123 +1,2 @@
 // Copyright (C) 2016-2026 Michael Kazakov. Subject to GNU General Public License version 3.
-#include "NCPanelPathBarPresentation.h"
-
-#include <Foundation/Foundation.h>
-
-namespace nc::panel {
-
-static void MarkCurrentDirectoryBreadcrumb(std::vector<PanelHeaderBreadcrumb> &breadcrumbs)
-{
-    if( !breadcrumbs.empty() )
-        breadcrumbs.back().is_current_directory = true;
-}
-
-[[nodiscard]] std::vector<PanelHeaderBreadcrumb> BuildPanelHeaderBreadcrumbs(const PanelPathContext &path_context)
-{
-    std::vector<PanelHeaderBreadcrumb> out;
-    const auto &verbose_full = path_context.verbose_full_path;
-    if( verbose_full.empty() )
-        return out;
-
-    // Must match VerboseDirectoryFullPath(), which appends '/' when Directory() omits it.
-    std::string dir_slash = path_context.directory_path;
-    if( dir_slash.empty() )
-        return out;
-    if( dir_slash.back() != '/' )
-        dir_slash += '/';
-    if( verbose_full.size() < dir_slash.size() )
-        return out;
-    if( !verbose_full.ends_with(dir_slash) )
-        return out;
-
-    std::string junction = verbose_full.substr(0, verbose_full.size() - dir_slash.size());
-    while( !junction.empty() && junction.back() == '/' )
-        junction.pop_back();
-
-    const std::string &path_only = path_context.posix_path;
-
-    if( !junction.empty() ) {
-        PanelHeaderBreadcrumb j;
-        j.label = [NSString stringWithUTF8String:junction.c_str()];
-        j.navigate_to_vfs_path = "/";
-        out.push_back(std::move(j));
-    }
-
-    if( path_only == "/" ) {
-        if( junction.empty() ) {
-            PanelHeaderBreadcrumb b;
-            b.label = @"/";
-            out.push_back(std::move(b));
-        }
-        MarkCurrentDirectoryBreadcrumb(out);
-        return out;
-    }
-
-    if( path_only.empty() || path_only.front() != '/' )
-        return out;
-
-    if( junction.empty() ) {
-        PanelHeaderBreadcrumb root;
-        root.label = @"/";
-        root.navigate_to_vfs_path = "/";
-        out.push_back(std::move(root));
-    }
-
-    const std::string rest = path_only.substr(1);
-    std::vector<std::string> comps;
-    size_t start = 0;
-    while( start <= rest.size() ) {
-        const size_t slash = rest.find('/', start);
-        if( slash == std::string::npos ) {
-            if( start < rest.size() )
-                comps.emplace_back(rest.substr(start));
-            break;
-        }
-        if( slash > start )
-            comps.emplace_back(rest.substr(start, slash - start));
-        start = slash + 1;
-    }
-
-    std::string acc;
-    for( size_t i = 0; i < comps.size(); ++i ) {
-        acc += "/";
-        acc += comps[i];
-        PanelHeaderBreadcrumb b;
-        b.label = [NSString stringWithUTF8String:comps[i].c_str()];
-        if( i + 1 < comps.size() )
-            b.navigate_to_vfs_path = acc;
-        out.push_back(std::move(b));
-    }
-    MarkCurrentDirectoryBreadcrumb(out);
-    return out;
-}
-
-std::string NormalizePanelHeaderPOSIXPathForActions(const std::string &path_without_trailing_slash)
-{
-    std::string p = path_without_trailing_slash;
-    if( p.empty() )
-        return "/";
-    if( p.front() != '/' )
-        p.insert(p.begin(), '/');
-    while( p.size() > 1 && p.back() == '/' )
-        p.pop_back();
-    return p;
-}
-
-std::optional<std::string> ResolvePanelBreadcrumbSegmentPOSIXForMenu(bool is_current_directory,
-                                                                     const std::optional<std::string> &navigate_to_vfs_path,
-                                                                     const std::optional<std::string> &fallback_posix_path,
-                                                                     const std::optional<std::string> &plain_path)
-{
-    if( is_current_directory ) {
-        if( fallback_posix_path.has_value() && !fallback_posix_path->empty() )
-            return fallback_posix_path;
-        return std::nullopt;
-    }
-    if( navigate_to_vfs_path.has_value() && !navigate_to_vfs_path->empty() )
-        return navigate_to_vfs_path;
-    if( fallback_posix_path.has_value() && !fallback_posix_path->empty() )
-        return fallback_posix_path;
-    return plain_path;
-}
-
-} // namespace nc::panel
+// Implementation moved to the Panel library (Source/Panel/source/PathBar.mm).
